requirements.txt
Pillow~=8.2.0
code128~=0.3
requests~=2.25.1
pyodbc~=4.0.30
telegram~=0.0.1
tabulate~=0.8.9
python-telegram-bot~=12.8
